[build] use freestanding headers

+++ include/drivers/screen.hpp
@@ -1,7 +1,7 @@
 #ifndef DOKKAN_DRIVERS_SCREEN_H
 #define DOKKAN_DRIVERS_SCREEN_H
 
-#include <cpu/types.hpp>
+#include <stdint.h>
 
 namespace dokkan::drivers {
 
@@ -9,7 +9,7 @@ class Screen {
  public:
   explicit Screen() = delete;
   static void print(const char* text);
-  static void printAt(const char* text, size_t col, size_t row);
+  static void printAt(const char* text, int16_t col, int16_t row);
   static void printLine(const char* text);
   static void printLine();
   static void clear();
